Pick real inventory hostname using a label
If kube_override_hostname is defined first_kube_control_plane fact will
return wrong value since the node names in the inventory will be
different that the node names in the cluster. Label the control plane
nodes with the label `kubespray.io/hostname` using `{{
inventory_hostname }}` and pick this if it is defined as the value of
first_kube_control_plane.